requirements.txt
# Core dependencies for AI-Viz-Lab
# Python 3.10+ required

# UI Framework
gradio>=4.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# ML/DL
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
sentencepiece>=0.1.99

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0

# Optional: For GPU support
# cuda-python>=12.0.0  # Uncomment if using NVIDIA GPU

# Optional: For advanced visualizations
# scikit-learn>=1.3.0  # For dimensionality reduction
# pandas>=2.0.0        # For data handling
